+++ src/App.tsx
@@ -4,6 +4,8 @@ import { Hero } from './components/sections/Hero';
 import { SynergyNarrative } from './components/sections/SynergyNarrative';
 import { ServicesBento } from './components/sections/ServicesBento';
 import { ExpertDirectory } from './components/sections/ExpertDirectory';
+import { FacilitiesBento } from './components/sections/FacilitiesBento';
+import { PortfolioSpotlight } from './components/sections/PortfolioSpotlight';
 
 export const App: React.FC = () => {
   return (
@@ -14,6 +16,8 @@ export const App: React.FC = () => {
         <SynergyNarrative />
         <ServicesBento />
         <ExpertDirectory />
+        <FacilitiesBento />
+        <PortfolioSpotlight />
       </main>
     </div>
   );